% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_eez_code}
\alias{get_eez_code}
\title{Function to pull numeric EEZ code using EEZ name}
\usage{
get_eez_code(eez_name, key)
}
\arguments{
\item{eez_name}{string, EEZ name}

\item{key}{string, API token}
}
\value{
dataframe, eez code and EEZ name for matching EEZs
}
\description{
Function to pull numeric EEZ code using EEZ name
}
