% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Function to get endpoint for given event type}
\usage{
get_endpoint(event_type = "port_visit", limit = 10, vessel)
}
\arguments{
\item{event_type}{Type of event to get data of. It can be "port_visit" or "fishing"}

\item{limit}{Number of events to import. We need some documentation for the max.}

\item{vessel}{VesselID. How to get this?}
}
\description{
Function to get endpoint for given event type
}
