% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Base function to get event from API and convert response to data frame}
\usage{
get_event(
  event_type = "port_visit",
  vessel = NULL,
  include_regions = NULL,
  start_date = NULL,
  end_date = NULL,
  confidences = NULL,
  key = gfw_auth(),
  quiet = FALSE
)
}
\arguments{
\item{event_type}{Type of event to get data of. One of "port_visit", "encounter", "loitering, or "fishing"}

\item{vessel}{VesselID. How to get this?}

\item{include_regions}{Whether to include regions? Ask engineering if this can always be false}

\item{start_date}{Start of date range to search events}

\item{end_date}{End of date range to search events}

\item{confidences}{Confidence levels (1-4) of events (port visits only).}

\item{key}{Authorization token. Can be obtained with gfw_auth function}

\item{quiet}{Boolean. Whether to print the number of events returned by the request.}
}
\description{
Base function to get event from API and convert response to data frame
}
\details{
There are currently four possible event types to request for a vessel. However, only likely vessels
can have fishing events (\code{event_type = 'fishing'}). An encounter event, which involves multiple vessels, will
be returned for each vessel involved in the encounter. For example, an encounter between a carrier vessel
and fishing vessel will be represented as two events, one for each vessel. These events will have identical \code{id}
values that are suffixed by a \code{.} and a number.
}
