% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_identity_endpoint}
\alias{get_identity_endpoint}
\title{Function to get API endpoint name for identity search}
\usage{
get_identity_endpoint(dataset_type, search_type, ...)
}
\arguments{
\item{dataset_type}{Type of identity dataset to get API dataset name for. It can be "support_vessel", "carrier_vessel" or "fishing_vessel"}

\item{search_type}{Type of vessel search to perform. Can be "basic", "advanced", or "id"}

\item{...}{Other arguments that would depend on the dataset type.}
}
\description{
Function to get API endpoint name for identity search
}
