% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_report.R
\name{get_last_report}
\alias{get_last_report}
\title{Base function to get status of last report generated}
\usage{
get_last_report(key = gfw_auth())
}
\arguments{
\item{key}{Authorization token. Can be obtained with \code{gfw_auth()} function}
}
\description{
Function to check the status of the last API request sent with get_raster().
}
\details{
The \code{get_last_report()} function will tell you if the APIs are still processing your request and
will download the results if the request has finished successfully. You will receive an error message
if the request finished but resulted in an error or if it's been >30 minutes since the last report was
generated using \code{get_raster()}.

For more information, see the https://globalfishingwatch.org/our-apis/documentation#get-last-report-generated.
}
\examples{
\dontrun{
get_last_report(key = gfw_auth())
}

}
