% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_regions}
\alias{get_regions}
\title{List of available regions}
\usage{
get_regions(region_source = "EEZ", key = gfw_auth())
}
\arguments{
\item{region_source}{string, source of region data ('eez', 'mpa', 'rfmo')}

\item{key}{string, API token}
}
\value{
dataframe, all region ids and names for specified region type
}
\description{
List of available regions
}
\examples{
get_regions(region_source = "EEZ")
get_regions(region_source = "RFMO")
get_regions(region_source = "MPA")
}
