% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_info.R
\name{get_vessel_info}
\alias{get_vessel_info}
\title{Base function to get vessel information from API and convert response to data frame}
\usage{
get_vessel_info(query = NULL, search_type = NULL, dataset = "all", key)
}
\arguments{
\item{query}{search terms to identify vessel}

\item{search_type}{type of search, may be 'basic','advanced', or 'id'}

\item{dataset}{identity datasets to search against, default = 'all'}
}
\description{
Base function to get vessel information from API and convert response to data frame
}
\details{
There are three search types. \code{basic} search takes features like MMSI,
IMO, callsign, shipname as inputs and identifies all vessels in the specified
dataset that match. \code{advanced} search allows for the use of fuzzy matching with
terms such as LIKE. The \code{id} search allows the user to search using a GFW vessel
id.
There are three identity databases available, \code{"carrier_vessel"}, \code{"support_vessel"},
and \code{"fishing_vessel"}. The user can also specify \code{"all"} and search again all databases
at once. This is generally recommended.
}
\examples{
get_vessel_info(query = 224224000, search_type = "basic")
get_vessel_info(query = "shipname LIKE '\%GABU REEFE\%' OR imo = '8300949'", search_type = "advanced")
get_vessel_info(query = "8c7304226-6c71-edbe-0b63-c246734b3c01", search_type = "id")
get_vessel_info(query = "8c7304226-6c71-edbe-0b63-c246734b3c01,6583c51e3-3626-5638-866a-f47c3bc7ef7c,71e7da672-2451-17da-b239-857831602eca", search_type = 'id')
}
