% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_track.R
\name{get_vessel_track}
\alias{get_vessel_track}
\title{Get vessel track}
\usage{
get_vessel_track(auth)
}
\arguments{
\item{auth}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Get vessel track
}
