% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_4wings.R
\name{gfw_4wings}
\alias{gfw_4wings}
\title{Retrieve datasets from the 4 wings API}
\usage{
gfw_4wings(
  api_endpoint = "AIS",
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  start_date = "2023-01-01",
  end_date = "2023-12-31",
  region_source = NULL,
  region = NULL,
  group_by = NULL,
  filter_by = NULL,
  key = gfw_auth(),
  print_request = FALSE
)
}
\arguments{
\item{api_endpoint}{API endpoint}

\item{spatial_resolution}{Raster spatial resolution. Can be \code{"LOW"} (0.1 degree)
or \code{"HIGH"} (0.01 degree).}

\item{temporal_resolution}{Raster temporal resolution. Can be \code{"HOURLY"},
\code{"DAILY"}, \code{"MONTHLY"}, \code{"YEARLY"}.}

\item{start_date}{Required. Start of date range to search events, in
YYYY-MM-DD format and including this date.}

\item{end_date}{Required. End of date range to search events, in
YYYY-MM-DD format and excluding this date.}

\item{region_source}{Required. Source of the region: \code{"EEZ"}, \code{"MPA"},
\code{"RFMO"} or \code{"USER_SHAPEFILE"}.}

\item{region}{Required. If \code{region_source} is set to \code{"EEZ"}, \code{"MPA"} or
\code{"RFMO"}, GFW region code (see \code{\link[=gfw_region_id]{gfw_region_id()}}). If
\code{region_source = "USER_SHAPEFILE"}, \code{sf} shapefile with the area of interest.}

\item{group_by}{Optional. Parameter to group by. Can be \code{"VESSEL_ID"}, \code{"FLAG"},
\code{"GEARTYPE"}, \code{"FLAGANDGEARTYPE"} or \code{"MMSI"}.}

\item{filter_by}{Fields to filter, variable depending on data source.}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}

\item{print_request}{Boolean. Whether to print the request, for debugging
purposes. When contacting the GFW team it will be useful to send this string.}
}
\description{
This is the basic internal function to retrieve 4wings datasets.
It is used under the hood by \code{gfw_ais_} and \code{gfw_sar_} functions, which are
split because the source data (AIS vs SAR) their documentation and parameters
differ
}
\details{
For detailed information about the 4Wings API, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#map-visualization-4wings-api
}

For more details on the 4Wings API data caveats, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#apparent-fishing-effort
\item https://globalfishingwatch.org/our-apis/documentation#sar-vessel-detections-data-caveats
\item https://globalfishingwatch.org/our-apis/documentation#ais-vessel-presence-caveats
}
}
\keyword{internal}
