% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gfw_api_request}
\alias{gfw_api_request}
\title{General function for GFW API requests, including handling of pagination.}
\usage{
gfw_api_request(endpoint, key)
}
\arguments{
\item{endpoint}{the endpoint to make the request}

\item{key}{Authentication key}
}
\description{
General function for GFW API requests, including handling of pagination.
}
