% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gfw_base_url}
\alias{gfw_base_url}
\title{Get the Global Fishing Watch API base URL}
\usage{
gfw_base_url()
}
\description{
Returns the \code{GFW_BASE_URL} environment variable if it is set and not empty
(ignoring leading/trailing whitespace). Otherwise, returns the default
production API base URL: \code{"https://gateway.api.globalfishingwatch.org/v3/"}.
}
