% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gfw_marine_regions}
\alias{gfw_marine_regions}
\title{Simplified Marine Regions v12 dataset (previously \code{marine_regions})}
\format{
A tibble with 285 rows and 5 columns
\describe{
\item{\code{iso}}{ISO Code derived from the different ISO code columns in the Maritime Boundaries Geodatabase v12}
\item{\code{name}}{Territory name derived from the different Name and Geoname columns in the Maritime Boundaries Geodatabase v12}
\item{\code{MRGID}}{Original numeric id for EEZs from the Marine Boundaries Geodatabase v12. This numeric code should be used in the \code{region} argument when using functions
\code{\link[=gfw_ais_fishing_hours]{gfw_ais_fishing_hours()}}, \code{\link[=gfw_ais_presence]{gfw_ais_presence()}}, and \code{\link[=gfw_sar_vessel_detections]{gfw_sar_vessel_detections()}} and \code{\link[=gfw_event]{gfw_event()}}, when \code{region_source = "EEZ"}.}
\item{\code{GEONAME}}{Original GEONAME field from the Maritime Boundaries Geodatabase v12}
\item{\code{POL_TYPE}}{Original POLygon TYPE from the Maritime Boundaries Geodatabase v12. Possible value are "200NM", "Overlapping claim" and "Joint regime"}
}
}
\source{
Marine Regions https://www.marineregions.org/. https://doi.org/10.14284/632
}
\usage{
gfw_marine_regions
}
\description{
A tibble with the ISO codes and names derived from
Marine Regions v12, to be used by \code{\link[=gfw_regions]{gfw_regions()}} and \code{\link[=gfw_region_id]{gfw_region_id()}} to
extract numeric EEZ codes (\code{MRGID})
}
\details{
The main source data for EEZ shapefiles and information across Global Fishing
Watch's data and platforms is Marine Regions v12.

The main variable of interest of this dataset is \code{MRGID}, the unique
identifier for each one of the polygons comprising Marine Regions. \code{MRGID}
is necessary to retrieve apparent fishing effort data and all event types
from other \code{gfwr} functions when \code{region_source = "EEZ"}

Functions \code{\link[=gfw_region_id]{gfw_region_id()}} and \code{\link[=gfw_regions]{gfw_regions()}} facilitate fetching region ids
from ISO codes and region names.

However, the original Marine Region dataset includes several columns that can
refer to the ISO of the territory itself (\code{ISO_TER1}), the entity that has
sovereignty over the territory (\code{ISO_SOV1}) or the territories that operate
Joint regime areas, or claim Overlapping claims areas (\code{ISO_SOV1}, \code{ISO_SOV2},
\code{ISO_SOV3}).

The table \code{gfw_marine_regions} unifies all these ISO markers into a single column without
modifying the original data, in a reproducible way.
\itemize{
\item Joint regime and overlapping claims with no ISO receive an \code{NA}: \code{iso = NA}
\item Territories within the 200NM with an ISO receive it: \code{iso = ISO_TER1}
\item Overlapping claims with an ISO receive it: \code{iso = ISO_TER1}
\item Territories within the 200NM and no ISO receive the ISO of their mainland: \code{iso = ISO_SOV1}
}

Likewise, \code{gfw_marine_regions} also derives a \code{name} column from the different
names and \code{GEONAME} columns present in the original dataset.
\itemize{
\item Countries and territories within the 200NM EEZs receive their name: \code{name = TERRITORY1}
\item Overlapping claims with a name receive their name: \code{name = TERRITORY1}
\item Joint regimes and overlapping claims with no name receive their \code{GEONAME}: \code{name = GEONAME}
}

The code implementing this synthesis can be found in the \code{data-raw/marine_regions.R} file.
}
\references{
Flanders Marine Institute (2023). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 12. Available online at https://www.marineregions.org/. https://doi.org/10.14284/632
}
\keyword{datasets}
