% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_region_id.R
\name{gfw_region_id}
\alias{gfw_region_id}
\title{Function to pull region code using region name and viceversa}
\usage{
gfw_region_id(
  region = NULL,
  region_source = "EEZ",
  key = gfw_auth(),
  region_name = deprecated()
)
}
\arguments{
\item{region}{Character or numeric EEZ MPA or RFMO name or id.}

\item{region_source}{Character, source of region data, \code{"EEZ"}, \code{"MPA"} or \code{"RFMO"}.}

\item{key}{Character, API token. Defaults to \code{gfw_auth()}.}

\item{region_name}{Deprecated, replaced by region.}
}
\value{
For \code{gfw_region_id()}, the corresponding code, region names or iso code
for the EEZ, MPA or RFMO label
}
\description{
Function to pull region code using region name and viceversa
}
\details{
For detailed information about the Regions API, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#regions
}

For more details on the Regions API data caveats, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#exclusive-economic-zone-boundaries-definitions
\item https://globalfishingwatch.org/our-apis/documentation#marine-protected-area-boundaries-definition
\item https://globalfishingwatch.org/our-apis/documentation#what-does-it-mean-if-an-event-is-within-a-specific-geographic-area-such-as-an-eez-mpa-or-rfmo
\item https://globalfishingwatch.org/our-apis/documentation#how-does-gfw-calculate-that-an-event-has-a-publicly-listed-authorization
\item https://globalfishingwatch.org/our-apis/documentation#insights-api-rfmo-iuu-vessel-list
}
}
\examples{
\dontrun{
gfw_region_id(region = "COL", region_source = "EEZ")
gfw_region_id(region = "Colombia", region_source = "EEZ")
gfw_region_id(region = "Nazca", region_source = "MPA")
gfw_region_id(region = "IOTC", region_source = "RFMO")
gfw_region_id(region = 8456, region_source = "EEZ")
# Handling empty strings (high-seas)
gfw_region_id(region = "", region_source = "EEZ")
gfw_region_id(region = NA, region_source = "EEZ")
gfw_region_id(region = NA, region_source = "MPA")
}
}
\seealso{
\code{\link[=gfw_regions]{gfw_regions()}}
}
