% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_region_id.R
\name{gfw_regions}
\alias{gfw_regions}
\title{List of available regions in Global Fishing Watch
platforms, EEZs, MPAs, and RFMOs}
\usage{
gfw_regions(region_source = "EEZ", key = gfw_auth())
}
\arguments{
\item{region_source}{string, source of region data ("EEZ", "MPA", "RFMO')}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}
}
\value{
A dataframe with all region ids and names for
specified region type
}
\description{
List of available regions in Global Fishing Watch
platforms, EEZs, MPAs, and RFMOs
}
\details{
For detailed information about the Regions API, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#regions
}

For more details on the Regions API data caveats, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#exclusive-economic-zone-boundaries-definitions
\item https://globalfishingwatch.org/our-apis/documentation#marine-protected-area-boundaries-definition
\item https://globalfishingwatch.org/our-apis/documentation#what-does-it-mean-if-an-event-is-within-a-specific-geographic-area-such-as-an-eez-mpa-or-rfmo
\item https://globalfishingwatch.org/our-apis/documentation#how-does-gfw-calculate-that-an-event-has-a-publicly-listed-authorization
\item https://globalfishingwatch.org/our-apis/documentation#insights-api-rfmo-iuu-vessel-list
}
}
\examples{
\dontrun{
gfw_regions(region_source = "EEZ")
gfw_regions(region_source = "RFMO")
gfw_regions(region_source = "MPA")
}
}
\seealso{
\code{\link[=gfw_region_id]{gfw_region_id()}}
}
