% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gfw_test_shape}
\alias{gfw_test_shape}
\title{A sample shapefile (previously \code{test_shape})}
\format{
A shapefile with a single polygon.
}
\usage{
gfw_test_shape
}
\description{
An sf shapefile to show as an example of user-defined GeoJSON in \code{\link[=gfw_event]{gfw_event()}},
\code{\link[=gfw_ais_fishing_hours]{gfw_ais_fishing_hours()}}, \code{\link[=gfw_ais_presence]{gfw_ais_presence()}}, and \code{\link[=gfw_sar_vessel_detections]{gfw_sar_vessel_detections()}}
}
\keyword{datasets}
