% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gist_error_body}
\alias{gist_error_body}
\title{Helper function to parse error message data
and display appropriately to user
Taken from httr2 docs: https://httr2.r-lib.org/articles/wrapping-apis.html#sending-data}
\usage{
gist_error_body(resp)
}
\value{

}
\description{
Helper function to parse error message data
and display appropriately to user
Taken from httr2 docs: https://httr2.r-lib.org/articles/wrapping-apis.html#sending-data
}
\keyword{internal}
