% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_char}
\alias{make_char}
\title{Basic function to make length 1 lists into characters}
\usage{
make_char(col)
}
\value{

}
\description{
Basic function to make length 1 lists into characters
}
\keyword{internal}
