% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_datetime}
\alias{make_datetime}
\title{Helper function to convert datetime responses}
\usage{
make_datetime(x)
}
\value{

}
\description{
Helper function to convert datetime responses
}
\keyword{internal}
