% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_to_geojson}
\alias{sf_to_geojson}
\title{Formats an sf shapefile to a formatted geojson}
\usage{
sf_to_geojson(sf_shape)
}
\arguments{
\item{sf_shape}{The sf shapefile to transform}
}
\value{
A correctly-formatted geojson to be used in \code{get_raster()}
}
\description{
Formats an sf shapefile to a formatted geojson
}
