% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_to_array}
\alias{vector_to_array}
\title{Transforms a vector to a named vector for httr2}
\usage{
vector_to_array(x, type = "vessel")
}
\arguments{
\item{x}{The vector to transform}

\item{type}{The type of data to paste, will be "events", "datasets", or "vessel" depending on the context}
}
\value{
A named vector in the format required by the API, with names followed
by a zero-indexed suffix (ex. datasets\[0\])
}
\description{
Transforms a vector to a named vector for httr2
}
\examples{
vector_to_array(x = 1, type = "vessel")
vector_to_array(x = "a", type = "vessel")
vector_to_array(x = c(1, 2), type = "dataset")
vector_to_array(x = c(1, 2, 3), type = "dataset")
vector_to_array(x = "fishing", type = "dataset")
vector_to_array(x = c("fishing", "port-visits"), type = "event")
}
