% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Function to get API dataset name for given event type}
\usage{
get_endpoint(dataset_type, ...)
}
\arguments{
\item{dataset_type}{Type of dataset to get API dataset name for. It can be "raster",
"ENCOUNTER", "LOITERING", "FISHING", "PORT_VISIT", "GAP", "EEZ", "RFMO" or "MPA"}

\item{...}{Other arguments that would depend on the dataset type.}
}
\description{
Function to get API dataset name for given event type
}
\keyword{internal}
