% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Base function to get event from API and convert response to data frame}
\usage{
get_event(
  event_type = "port_visit",
  vessel = NULL,
  include_regions = NULL,
  start_date = NULL,
  end_date = NULL,
  confidences = NULL,
  limit = 99999,
  offset = 0,
  key = gfw_auth()
)
}
\arguments{
\item{event_type}{Type of event to get data of. It can be "port_visit" or "fishing"}

\item{vessel}{VesselID. How to get this?}

\item{include_regions}{Whether to include regions? Ask engineering if this can always be false}

\item{start_date}{Start of date range to search events}

\item{end_date}{End of date range to search events}

\item{confidences}{Confidence levels (1-4) of events (port visits only).}

\item{limit}{Limit of response size for each GFW API call.}

\item{offset}{Internal parameter to GFW pagination}

\item{key}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get event from API and convert response to data frame
}
