% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Base function to get event from API and convert response to data frame}
\usage{
get_event(
  event_type = "port_visit",
  limit = 10,
  vessel = "6583c51e3-3626-5638-866a-f47c3bc7ef7c",
  auth
)
}
\arguments{
\item{event_type}{Type of event to get data of. It can be "port_visit" or "fishing"}

\item{limit}{Number of events to import. We need some documentation for the max.}

\item{vessel}{VesselID. How to get this?}

\item{auth}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get event from API and convert response to data frame
}
