% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Base function to get event from API and convert response to data frame}
\usage{
get_event(
  event_type = "port_visit",
  vessel = NULL,
  include_regions = NULL,
  start_date = NULL,
  end_date = NULL,
  key = gfw_auth()
)
}
\arguments{
\item{event_type}{Type of event to get data of. It can be "port_visit" or "fishing"}

\item{vessel}{VesselID. How to get this?}

\item{include_regions}{Whether to include regions? Ask engineering if this can always be false}

\item{start_date}{Start of date range to search events}

\item{end_date}{End of date range to search events}

\item{key}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get event from API and convert response to data frame
}
