% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_identity_endpoint}
\alias{get_identity_endpoint}
\title{Function to get API endpoint name for identity search}
\usage{
get_identity_endpoint(dataset_type, search_type, ids, ...)
}
\arguments{
\item{dataset_type}{Type of identity dataset to get API dataset name for. It can be a vector with any combination of "support_vessel", "carrier_vessel" or "fishing_vessel"}

\item{search_type}{Type of vessel search to perform. Can be "search" or "id". "advanced" is no longer in use as of gfwr 2.0.0 and basic and advanced options can be accessed with parameters query and where}

\item{ids}{optional, a vector with vessel ids}

\item{...}{Other arguments that would depend on the dataset type.}
}
\description{
Function to get API endpoint name for identity search
}
