% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster.R
\name{get_raster}
\alias{get_raster}
\title{Base function to get raster from API and convert response to data frame}
\usage{
get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  group_by = NULL,
  start_date = NULL,
  end_date = NULL,
  format = "csv",
  shape_json = NULL,
  key
)
}
\arguments{
\item{spatial_resolution}{raster spatial resolution. Can be "low" = 0.1 degree or "high" = 0.01 degree}

\item{temporal_resolution}{raster temporal resolution. Can be 'daily','monthly','yearly'.}

\item{group_by}{parameter to group by. Can be 'vessel_id', 'flag', 'geartype', 'flagAndGearType'}

\item{start_date}{Start of date range for raster}

\item{end_date}{End of date range for raster}

\item{format}{output format. Current support for 'csv'.}

\item{shape_json}{geojson, shape to filter raster}

\item{key}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get raster from API and convert response to data frame
}
