% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster.R
\name{get_raster}
\alias{get_raster}
\title{Base function to get raster from API and convert response to data frame}
\usage{
get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  group_by = NULL,
  filter_by = NULL,
  date_range = NULL,
  region = NULL,
  region_source = NULL,
  key = gfw_auth()
)
}
\arguments{
\item{spatial_resolution}{raster spatial resolution. Can be "LOW" = 0.1 degree
or "HIGH" = 0.01 degree}

\item{temporal_resolution}{raster temporal resolution. Can be 'HOURLY',
'DAILY', 'MONTHLY', 'YEARLY'}

\item{group_by}{parameter to group by. Can be 'VESSEL_ID', 'FLAG', 'GEARTYPE',
'FLAGANDGEARTYPE' or 'MMSI'}

\item{filter_by}{parameter to filter by.}

\item{date_range}{Start and end of date range for raster (must be 366 days or less)}

\item{region}{geojson shape to filter raster or GFW region code (such as an
EEZ code). See details about formatting the geojson}

\item{region_source}{source of the region ('EEZ','MPA', 'RFMO' or 'USER_JSON')}

\item{key}{Authorization token. Can be obtained with gfw_auth() function}
}
\description{
Base function to get raster from API and convert response to data frame
}
\details{
The user-defined geojson has to be surrounded by a geojson tag,
that can be created using a simple paste:

\if{html}{\out{<div class="sourceCode">}}\preformatted{geojson_tagged <- paste0('\{"geojson":', your_geojson,'\}').
}\if{html}{\out{</div>}}

If you have an \strong{sf} shapefile, you can also use function \code{\link[=sf_to_geojson]{sf_to_geojson()}}
to obtain the correctly-formatted geojson.
}
