% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster.R
\name{get_raster}
\alias{get_raster}
\title{Base function to get raster from API and convert response to data frame}
\usage{
get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  group_by = NULL,
  date_range = NULL,
  region = NULL,
  region_source = NULL,
  key = gfw_auth()
)
}
\arguments{
\item{spatial_resolution}{raster spatial resolution. Can be "low" = 0.1 degree or "high" = 0.01 degree}

\item{temporal_resolution}{raster temporal resolution. Can be 'daily','monthly','yearly'.}

\item{group_by}{parameter to group by. Can be 'vessel_id', 'flag', 'gearType', 'flagAndGearType'}

\item{date_range}{Start and end of date range for raster}

\item{region}{geojson or GFW region code, shape to filter raster}

\item{region_source}{source of the region ('eez','mpa', 'trfmo' or 'user_json')}

\item{key}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get raster from API and convert response to data frame
}
\details{
See examples at https://github.com/GlobalFishingWatch/gfwr
}
