% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_id.R
\name{get_region_id}
\alias{get_region_id}
\title{Function to pull region code using region name and viceversa}
\usage{
get_region_id(region_name = NULL, region_source = "EEZ", key = gfw_auth())
}
\arguments{
\item{region_name}{Character or numeric EEZ MPA or RFMO name or id.}

\item{region_source}{Character, source of region data, \code{"EEZ"}, \code{"MPA"} or \code{"RFMO"}.}

\item{key}{Character, API token. Defaults to \code{gfw_auth()}.}
}
\value{
For \code{get_region_id()}, the corresponding code, region names or iso code
for the EEZ, MPA or RFMO label
}
\description{
Function to pull region code using region name and viceversa
}
\examples{
\dontrun{
get_region_id(region_name = "COL", region_source = "EEZ")
get_region_id(region_name = "Colombia", region_source = "EEZ")
get_region_id(region_name = "Nazca", region_source = "MPA")
get_region_id(region_name = "IOTC", region_source = "RFMO")
get_region_id(region_name = 8456, region_source = "EEZ")
# Handling empty strings (high-seas)
get_region_id(region_name = "", region_source = "EEZ")
get_region_id(region_name = NA, region_source = "EEZ")
get_region_id(region_name = NA, region_source = "MPA")
}
}
\seealso{
\code{\link[=get_regions]{get_regions()}}
}
