% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_info.R
\name{get_vessel_info}
\alias{get_vessel_info}
\title{Base function to get vessel information from API and convert response to data frame}
\usage{
get_vessel_info(
  query = NULL,
  search_type = NULL,
  dataset = "all",
  limit = 10000,
  offset = 0,
  key = gfw_auth()
)
}
\arguments{
\item{query}{search terms to identify vessel}

\item{search_type}{type of search, may be 'basic','advanced', or 'id'}

\item{dataset}{identity datasets to search against, default = 'all'}

\item{limit}{max number of entries to return in each API response. All results will
be returned regardless of limit}

\item{key}{Authorization token. Can be obtained with gfw_auth function}
}
\description{
Base function to get vessel information from API and convert response to data frame
}
\details{
There are three search types. \code{basic} search takes features like MMSI,
IMO, callsign, shipname as inputs and identifies all vessels in the specified
dataset that match. \code{advanced} search allows for the use of fuzzy matching with
terms such as LIKE. The \code{id} search allows the user to search using a GFW vessel
id.
There are three identity databases available, \code{"carrier_vessel"}, \code{"support_vessel"},
and \code{"fishing_vessel"}. The user can also specify \code{"all"} and search again all databases
at once. This is generally recommended.
}
