% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_ais_presence.R
\name{gfw_ais_presence}
\alias{gfw_ais_presence}
\title{Retrieve vessel presence from AIS data and convert response to tibble}
\usage{
gfw_ais_presence(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  start_date = NULL,
  end_date = NULL,
  region_source = NULL,
  region = NULL,
  group_by = NULL,
  filter_by = NULL,
  key = gfw_auth(),
  print_request = FALSE
)
}
\arguments{
\item{spatial_resolution}{Raster spatial resolution. Can be \code{"LOW"} (0.1 degree)
or \code{"HIGH"} (0.01 degree).}

\item{temporal_resolution}{Raster temporal resolution. Can be \code{"HOURLY"},
\code{"DAILY"}, \code{"MONTHLY"}, \code{"YEARLY"}.}

\item{start_date}{Required. Start of date range to search events, in
YYYY-MM-DD format and including this date.}

\item{end_date}{Required. End of date range to search events, in
YYYY-MM-DD format and \strong{excluding this date}.}

\item{region_source}{Required. Source of the region: \code{"EEZ"}, \code{"MPA"},
\code{"RFMO"} or \code{"USER_SHAPEFILE"}.}

\item{region}{Required. If \code{region_source} is set to \code{"EEZ"}, \code{"MPA"} or
\code{"RFMO"}, GFW region code (see \code{\link[=gfw_region_id]{gfw_region_id()}}). If
\code{region_source = "USER_SHAPEFILE"}, \code{sf} shapefile with the area of interest.}

\item{group_by}{Optional. Parameter to group by. Can be \code{"VESSEL_ID"}, \code{"FLAG"},
\code{"GEARTYPE"}, \code{"FLAGANDGEARTYPE"} or \code{"MMSI"}.}

\item{filter_by}{Fields to filter AIS-based vessel presence. Possible options
are \code{flag}, \code{vessel_type}, \code{speed}. See Details for more information}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}

\item{print_request}{Boolean. Whether to print the request, for debugging
purposes. When contacting the GFW team it will be useful to send this string.}
}
\description{
This function communicates with the \code{public-global-presence} latest dataset to
extract global vessel presence derived from AIS data. The presence is
determined by taking one position per hour per vessel from the positions
transmitted by the vessel's AIS. Unlike the apparent fishing effort dataset,
this includes all vessel types and focuses on presence rather than fishing
classification.
}
\details{
The \code{filter_by} parameter accepts \code{flag}, \code{vessel_type}, \code{speed} options.
This parameter accepts SQL expressions like \code{filter_by = "flag IN ('ESP')"},
\verb{flag in ('ESP', 'USA')}, \code{vessel_type = 'cargo'}. Accepted vessel speed
ranges are the following:
\itemize{
\item \verb{<2} – Less than 2 knots
\item \code{2-4} – 2 to 4 knots
\item \code{4-6} – 4 to 6 knots
\item \code{6-10} – 6 to 10 knots
\item \code{10-15} – 10 to 15 knots
\item \code{15-25} – 15 to 25 knots
\item \verb{>25} – Greater than 25 knots
}

For detailed information about the 4Wings AIS vessel presence API, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#ais-vessel-presence
}

For more details on the 4Wings AIS vessel presence API data caveats, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#ais-vessel-presence-caveats
}
}
\examples{
\dontrun{
library(gfwr)
# using region codes

code_eez <- gfw_region_id(region_name = "CIV", region_source = "EEZ")
gfw_ais_presence(spatial_resolution = "LOW",
           temporal_resolution = "YEARLY",
           group_by = "FLAG",
           start_date = "2021-01-01",
           end_date = "2021-01-10",
           region = code_eez$id,
           region_source = "EEZ",
           key = gfw_auth(),
           print_request = TRUE)

# Grouping by flag

code_mpa <- gfw_region_id(region_name = "Galapagos", region_source = "MPA")
gfw_ais_presence(spatial_resolution = "LOW",
           temporal_resolution = "MONTHLY",
           group_by = "FLAG",
           start_date = "2022-01-01",
           end_date = "2022-02-01",
           region = code_mpa$id[3],
           region_source = "MPA")

# Filter by flag but grouping by VESSEL_ID

code_mpa <- gfw_region_id(region_name = "Galapagos", region_source = "MPA")
gfw_ais_presence(spatial_resolution = "LOW",
           temporal_resolution = "MONTHLY",
           group_by = "VESSEL_ID",
           filter_by = "flag in ('ECU')",
           start_date = "2022-01-01",
           end_date = "2022-02-01",
           region = code_mpa$id[3],
           region_source = "MPA")

code_rfmo <- gfw_region_id(region_name = "GFCM", region_source = "RFMO")
gfw_ais_presence(spatial_resolution = "LOW",
           temporal_resolution = "YEARLY",
           start_date = "2022-01-01",
           end_date = "2023-01-01",
           region = code_rfmo$id[1],
           region_source = "RFMO")

# Using a sf from disk /loading a test sf object
data(test_shape)
gfw_ais_presence(spatial_resolution = "LOW",
            temporal_resolution = "YEARLY",
            start_date = "2021-01-01",
            end_date = "2021-02-01",
            region = test_shape,
            region_source = "USER_SHAPEFILE",
            key = gfw_auth(),
            print_request = TRUE)
}
}
\references{
AIS vessel presence examples in the API documentation https://globalfishingwatch.org/our-apis/documentation#report-ais-vessel-presence-examples
}
