% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_event_stats.R
\name{gfw_event_stats}
\alias{gfw_event_stats}
\title{Get events stats from API and convert response to tibble}
\usage{
gfw_event_stats(
  event_type,
  start_date = "2012-01-01",
  end_date = "2024-12-31",
  interval = NULL,
  vessels = NULL,
  flags = NULL,
  vessel_types = NULL,
  region_source = NULL,
  region = NULL,
  duration = 1,
  encounter_types = NULL,
  confidences = c(2, 3, 4),
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)
}
\arguments{
\item{event_type}{Type of event to get data of. A vector with any combination
of "ENCOUNTER", "FISHING", "GAP", "LOITERING", "PORT_VISIT"}

\item{start_date}{Start of date range to search events, in YYYY-MM-DD format
and including this date}

\item{end_date}{End of date range to search events, in YYYY-MM-DD format and
\strong{excluding this date}}

\item{interval}{Time series granularity. Must be a string. Possible values: 'HOUR', 'DAY', 'MONTH', 'YEAR'.}

\item{vessels}{A vector of vesselIds, obtained via \code{\link[=gfw_vessel_info]{gfw_vessel_info()}}}

\item{flags}{ISO3 code for the flag of the vessels. Null by default.}

\item{vessel_types}{A vector of vessel types, any combination of: \code{"FISHING"},
\code{"CARRIER"}, \code{"SUPPORT"}, \code{"PASSENGER"}, \code{"OTHER_NON_FISHING"}, \code{"SEISMIC_VESSEL"},
\code{"BUNKER_OR_TANKER"}, \code{"CARGO"}}

\item{region_source}{Optional. Source of the region ('EEZ','MPA', 'RFMO' or
'USER_SHAPEFILE').}

\item{region}{Optional but required if a value for \code{region_source} is specified.
If \code{region_source} is set to "EEZ", "MPA" or "RFMO", GFW region
code (see \code{\link[=gfw_region_id]{gfw_region_id()}}). If \code{region_source = "USER_SHAPEFILE"}, \code{sf}
shapefile with the area of interest.}

\item{duration}{minimum duration of the event in minutes. The default value is 1.}

\item{encounter_types}{Only useful when \code{event_type = "ENCOUNTER"}. Filters for
types of vessels during the encounter. A
vector with any combination of: \code{"CARRIER-FISHING"}, \code{"FISHING-CARRIER"},
\code{"FISHING-SUPPORT"}, \code{"SUPPORT-FISHING"}.}

\item{confidences}{Only useful when event_type = "PORT_VISIT". Confidence
levels (2-4) of events.}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}

\item{quiet}{Boolean. Whether to print the number of events returned by the
request}

\item{print_request}{Boolean. Whether to print the request, for debugging
purposes. When contacting the GFW team it will be useful to send this string}

\item{...}{Other arguments}
}
\description{
Get events stats from API and convert response to tibble
}
\details{
There are currently four available event types and these events are provided
for three vessel types - fishing, carrier, and support vessels.
Fishing events (\code{event_type = "FISHING"}) are specific to fishing vessels and
loitering events (\code{event_type = "LOITERING"}) are specific to carrier vessels.
Port visits (\code{event_type = "PORT_VISIT"}) and encounters
(\code{event_type = "ENCOUNTER"}) are available for all vessel types. For more
details about the various event types, see the
\href{https://globalfishingwatch.org/our-apis/documentation#data-caveat}{GFW API documentation}.
}
\examples{
\dontrun{
library(gfwr)
 # stats for encounters involving Russian carriers in given time range
gfw_event_stats(event_type = 'ENCOUNTER',
encounter_types = c("CARRIER-FISHING","FISHING-CARRIER"),
vessel_types = 'CARRIER',
start_date = "2018-01-01",
end_date = "2023-01-31",
flags = 'RUS',
duration = 60,
interval = "YEAR")
 # port visits stats in a region (Senegal)
 gfw_event_stats(event_type = 'PORT_VISIT',
start_date = "2018-01-01",
end_date = "2019-01-31",
confidences = c('3','4'),
region = 8371,
region_source = 'EEZ',
interval = "YEAR")
}
}
