% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_region_id.R
\name{gfw_regions}
\alias{gfw_regions}
\title{List of available regions in Global Fishing Watch
platforms, EEZs, MPAs, and RFMOs}
\usage{
gfw_regions(region_source = "EEZ", key = gfw_auth())
}
\arguments{
\item{region_source}{string, source of region data ("EEZ", "MPA", "RFMO')}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}
}
\value{
A dataframe with all region ids and names for
specified region type
}
\description{
List of available regions in Global Fishing Watch
platforms, EEZs, MPAs, and RFMOs
}
\examples{
\dontrun{
gfw_regions(region_source = "EEZ")
gfw_regions(region_source = "RFMO")
gfw_regions(region_source = "MPA")
}
}
\seealso{
\code{\link[=gfw_region_id]{gfw_region_id()}}
}
