% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_renamed.R
\name{get_raster}
\alias{get_raster}
\alias{get_event}
\alias{get_event_stats}
\alias{get_region_id}
\alias{get_regions}
\alias{get_vessel_info}
\alias{get_last_report}
\alias{marine_regions}
\alias{test_shape}
\title{Functions renamed in gfwr 3.0}
\usage{
get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  start_date = "2023-01-01",
  end_date = "2023-12-31",
  region_source = NULL,
  region = NULL,
  group_by = NULL,
  filter_by = NULL,
  key = gfw_auth(),
  print_request = FALSE
)

get_event(
  event_type,
  start_date = "2012-01-01",
  end_date = "2024-12-31",
  sort = "+start",
  vessels = NULL,
  flags = NULL,
  vessel_types = NULL,
  region_source = NULL,
  region = NULL,
  duration = 1,
  encounter_types = NULL,
  gap_intentional_disabling = deprecated(),
  confidences = c(2, 3, 4),
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)

get_event_stats(
  event_type,
  start_date = "2012-01-01",
  end_date = "2024-12-31",
  interval = NULL,
  vessels = NULL,
  flags = NULL,
  vessel_types = NULL,
  region_source = NULL,
  region = NULL,
  duration = 1,
  encounter_types = NULL,
  confidences = c(2, 3, 4),
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)

get_region_id(region_name = NULL, region_source = "EEZ", key = gfw_auth())

get_regions(region_source = "EEZ", key = gfw_auth())

get_vessel_info(
  query = NULL,
  where = NULL,
  search_type = "search",
  ids = NULL,
  includes = c("AUTHORIZATIONS", "OWNERSHIP", "MATCH_CRITERIA"),
  match_fields = NULL,
  registries_info_data = c("ALL"),
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)

get_last_report(key = gfw_auth())

marine_regions()

test_shape()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Several functions were renamed in gwfr v 3.0 for more descriptive names and a common \code{gfw_} suffix
\itemize{
\item \code{get_raster()}      -> \code{gfw_ais_fishing_hours()}
\item \code{get_vessel_info()} -> \code{gfw_vessel_info()}
\item \code{get_event()}       -> \code{gfw_event()}
\item \code{get_event_stats()} -> \code{gfw_event_stats()}
\item \code{get_regions()}     -> \code{gfw_regions()}
\item \code{get_region_id()}   -> \code{gfw_region_id()}
\item \code{get_last_report()} -> \code{gfw_last_report()}
}

The two datasets were also renamed
\itemize{
\item \code{marine_regions} -> \code{gfw_marine_regions}
\item \code{test_shape} -> \code{gfw_test_shape}
}
}
\keyword{internal}
