% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_sar_vessel_detections.R
\name{gfw_sar_vessel_detections}
\alias{gfw_sar_vessel_detections}
\title{Retrieve vessel presence detected using SAR and convert response to tibble}
\usage{
gfw_sar_vessel_detections(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  start_date = NULL,
  end_date = NULL,
  region_source = NULL,
  region = NULL,
  group_by = NULL,
  filter_by = NULL,
  key = gfw_auth(),
  print_request = FALSE
)
}
\arguments{
\item{spatial_resolution}{Raster spatial resolution. Can be \code{"LOW"} (0.1 degree)
or \code{"HIGH"} (0.01 degree).}

\item{temporal_resolution}{Raster temporal resolution. Can be \code{"HOURLY"},
\code{"DAILY"}, \code{"MONTHLY"}, \code{"YEARLY"}.}

\item{start_date}{Required. Start of date range to search events, in
YYYY-MM-DD format and including this date.}

\item{end_date}{Required. End of date range to search events, in
YYYY-MM-DD format and excluding this date.}

\item{region_source}{Required. Source of the region: \code{"EEZ"}, \code{"MPA"},
\code{"RFMO"} or \code{"USER_SHAPEFILE"}.}

\item{region}{Required. If \code{region_source} is set to \code{"EEZ"}, \code{"MPA"} or
\code{"RFMO"}, GFW region code (see \code{\link[=gfw_region_id]{gfw_region_id()}}). If
\code{region_source = "USER_SHAPEFILE"}, \code{sf} shapefile with the area of interest.}

\item{group_by}{Optional. Parameter to group by. Can be \code{"VESSEL_ID"}, \code{"FLAG"},
\code{"GEARTYPE"}, \code{"FLAGANDGEARTYPE"} or \code{"MMSI"}.}

\item{filter_by}{Fields to filter SAR vessel detections. See Details for possible options.}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}

\item{print_request}{Boolean. Whether to print the request, for debugging
purposes. When contacting the GFW team it will be useful to send this string.}
}
\description{
Retrieve vessel presence detected using SAR and convert response to tibble
}
\details{
Possible filter options are:
\itemize{
\item \code{matched} – Whether detection matched with AIS data Example: \code{"matched='true'"} or \code{"matched='false'"}
\item flag – Vessel flag state (when matched with AIS). Example: \code{"flag in ('ESP', 'USA')"}
\item \code{vessel_id} – Vessel identifier (when matched with AIS). See the identity vignette for details about Vessel ID.
\item \code{geartype} – Fishing gear type (when matched with AIS) → View \href{https://globalfishingwatch.org/our-apis/documentation#gear-types-supported}{supported gear types}. Example: \code{"geartype in ('tuna_purse_seines', 'driftnets')"}
\item \code{neural_vessel_type} – AI classification based on neural network model. Values: <= 0.1: "Likely non-fishing", >= 0.9: "Likely fishing", 0.1 - 0.9: "Other/Unknown"
\item \code{shiptype} – Vessel type classification (when matched with AIS) → See \href{https://globalfishingwatch.org/our-apis/documentation#vessel-types}{Vessel types}
}
}
\examples{
\dontrun{
library(gfwr)
# using region codes
code_eez <- gfw_region_id(region = "Chile", region_source = "EEZ")
gfw_sar_vessel_detections(spatial_resolution = "LOW",
           temporal_resolution = "YEARLY",
           group_by = "FLAG",
           start_date = "2021-01-01",
           end_date = "2022-01-01",
           region = code_eez$id,
           region_source = "EEZ",
           key = gfw_auth())
# filter by matched
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          group_by = "VESSEL_ID",
                          filter_by = "matched = 'true'",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
## Unmatched vessels will have no id information:
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          group_by = "VESSEL_ID",
                          filter_by = "matched = 'false'",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
# Filter by flag
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          group_by = "VESSEL_ID",
                          filter_by = "flag IN ('PER', 'ECU')",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
# Filter by vessel ID
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          filter_by = "vessel_id = '320335fcf-fbe5-54e0-9367-b36ae25b64b5'",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
# Filter by geartype
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          group_by = "VESSEL_ID",
                          filter_by = "geartype IN ('tuna_purse_seines',
                           'driftnets')",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
# Filter by neural_vessel_type
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          filter_by = "neural_vessel_type = '0.3'",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())
# Filter by shiptype (vessel type)
gfw_sar_vessel_detections(spatial_resolution = "LOW",
                          temporal_resolution = "YEARLY",
                          filter_by = "shiptype = 'CARGO'",
                          start_date = "2021-01-01",
                          end_date = "2022-01-01",
                          region = code_eez$id,
                          region_source = "EEZ",
                          key = gfw_auth())


}
}
