% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_vessel_info.R
\name{gfw_vessel_info}
\alias{gfw_vessel_info}
\title{Base function to get vessel information from API and convert response to tibble}
\usage{
gfw_vessel_info(
  query = NULL,
  where = NULL,
  search_type = "search",
  ids = NULL,
  includes = NULL,
  match_fields = NULL,
  registries_info_data = c("ALL"),
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)
}
\arguments{
\item{query}{When \code{search_type = "search"}, a length-1 vector with the identity
variable of interest, MMSI, IMO, call sign or ship name. A search by vessel
name will return fuzzy results to account for variation in vessel names and
potential misspellings}

\item{where}{When \code{search_type = "search"}, an SQL expression to find the vessel of interest.}

\item{search_type}{Type of vessel search to perform. Can be \code{"search"} (the
default) to search for identity markers or \code{"id"} to search per \code{vesselId}.
(Note:\code{"advanced"} and \code{"basic"} are no longer in use as of \verb{gfwr 2.0.0.}).}

\item{ids}{When \code{search_type = "id"}, a vector with one or more \code{vesselId}s
of interest.}

\item{includes}{Optional. Enhances the response with additional
information depending on the selected \code{search_type}. If not specified, all
supported values for the selected \code{search_type} will be returned. See
\strong{Details} below}

\item{match_fields}{Optional. Allows to filter by \code{matchFields} levels.
Possible values: \code{"SEVERAL_FIELDS"}, \code{"NO_MATCH"}, \code{"ALL"}. Incompatible with \code{where}}

\item{registries_info_data}{when \code{search_type == "id"}, gets all the registry
objects, only the delta or the latest.
\describe{
\item{\code{"NONE"}}{The API will return the most recent object only}
\item{\code{"DELTA"}}{The API will return only the objects when the vessel
changed one or more identity properties}
\item{\code{"ALL"}}{The \code{registryInfo} array will return all objects we have in the vessel database}
}}

\item{key}{Character, API token. Defaults to \code{\link[=gfw_auth]{gfw_auth()}}.}

\item{quiet}{Boolean. Whether to print the number of events returned by the
request and progress. Default is FALSE.}

\item{print_request}{Boolean. Whether to print the request, for debugging
purposes. When contacting the GFW team it will be useful to send this string.}

\item{...}{Other parameters, see API documentation.}
}
\description{
Base function to get vessel information from API and convert response to tibble
}
\details{
\itemize{
\item When \code{search_type = "search"} the search takes basic identity features like
MMSI, IMO, callsign, shipname as inputs, using parameter \code{"query"}. For more advanced
SQL searches, use parameter \code{"where"}. You can combine logic operators like \code{AND},
\code{OR}, \code{=}, \code{>=} , <, \code{LIKE} (for fuzzy matching).
\item Parameter \strong{\code{includes}}: When \code{search_type = "search"}, supported values are:
\describe{
\item{\code{"OWNERSHIP"}}{returns ownership information}
\item{\code{"AUTHORIZATIONS"}}{lists public authorizations for the vessel}
\item{\code{"MATCH_CRITERIA"}}{adds information about the reason why a vessel is
returned. This provides related \code{vesselId}s identified through
matching with vessel registry records and represents Global Fishing Watch's
best estimate for linking AIS (self-reported) vessel positions to Vessel
Identity information derived from public registries.}
}
When \code{search_type = "id"}, the supported value is
\code{"POTENTIAL_RELATED_SELF_REPORTED_INFO"} and will returns all potential
related self-reported vessel information mentioned above.
}

For detailed information about the Vessels API, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#vessels-api
}

For more details on the Vessels API data caveats, please refer to the official
Global Fishing Watch API documentation:
\itemize{
\item https://globalfishingwatch.org/our-apis/documentation#vessel-api-vessel-identity-information
\item https://globalfishingwatch.org/our-apis/documentation#api-dataset
}
}
\examples{
\dontrun{
library(gfwr)

# Simple search

gfw_vessel_info(query = 224224000, search_type = "search")

# Advanced search with where instead of query:
gfw_vessel_info(where = "ssvid = '441618000' OR imo = '9047271'",
                search_type = "search")

 # Vessel id search

 gfw_vessel_info(search_type = "id",
 ids = c("8c7304226-6c71-edbe-0b63-c246734b3c01",
 "6583c51e3-3626-5638-866a-f47c3bc7ef7c"))
 }
}
\references{
Park, J., Van Osdel, J., Turner, J., Farthing, C.M., Miller, N.A.,
Linder, H.L., Ortuño Crespo, G., Carmine, G., Kroodsma, D.A., 2023. Tracking
elusive and shifting identities of the global fishing fleet. Science Advances
9, eabp8200. \url{https://doi.org/10.1126/sciadv.abp8200}
}
\seealso{
For more details check the \href{https://globalfishingwatch.github.io/gfwr/articles/identity.html}{Vessel identity vignette}

See also how the Vessel API is used in \href{https://globalfishingwatch.org/our-apis/assets/2024_Vessel_Viewer_and_APIs_behind_It.pdf}{Vessel Viewer}
}
