% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfwr-package.R
\docType{package}
\name{gfwr-package}
\alias{gfwr}
\alias{gfwr-package}
\title{gfwr: Interactions with GFW APIs providing data tibbles}
\description{
This package connects to several Global Fishing Watch APIs to get vessel and events information in an R-friendly format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/gfwr}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/gfwr/issues}
}

}
\author{
\strong{Maintainer}: Tyler Clavelle \email{tyler@globalfishingwatch.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Rocío Joo \email{rocio.joo@globalfishingwatch.org} (\href{https://orcid.org/0000-0003-0319-4210}{ORCID})
  \item Nate Miller \email{nate@globalfishingwatch.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
