% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfwr-package.R
\docType{package}
\name{gfwr-package}
\alias{gfwr}
\alias{gfwr-package}
\title{gfwr: Access data from Global Fishing Watch APIs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package connects to several Global Fishing Watch APIs to get vessel and events information in an R-friendly format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/gfwr}
  \item \url{https://globalfishingwatch.github.io/gfwr/}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/gfwr/issues}
}

}
\author{
\strong{Maintainer}: Andrea Sánchez-Tapia \email{andrea.sanchez-tapia@globalfishingwatch.org} (\href{https://orcid.org/0000-0002-3521-4338}{ORCID})

Authors:
\itemize{
  \item Tyler Clavelle \email{tyler@globalfishingwatch.org} (\href{https://orcid.org/0000-0002-2155-2459}{ORCID})
  \item Rocío Joo \email{rocio.joo@globalfishingwatch.org} (\href{https://orcid.org/0000-0003-0319-4210}{ORCID})
  \item Nate Miller \email{nate@globalfishingwatch.org}
  \item Jorge Cornejo \email{jorge.cornejo@globalfishingwatch.org}
}

}
\keyword{internal}
