% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_to_geojson}
\alias{sf_to_geojson}
\title{Formats an sf shapefile to a formatted geojson}
\usage{
sf_to_geojson(sf_shape, endpoint = "raster")
}
\arguments{
\item{sf_shape}{The sf shapefile to transform}

\item{endpoint}{The GFW endpoint destination for the geojson ("raster" or "event")}
}
\value{
A correctly-formatted geojson to be used in \code{\link[=get_raster]{get_raster()}} or \code{\link[=get_event]{get_event()}}
}
\description{
Formats an sf shapefile to a formatted geojson
}
\examples{
library(gfwr)
data(test_shape)
substr(sf_to_geojson(test_shape), 1, 20)
substr(sf_to_geojson(test_shape, endpoint = "event"), 1, 20)
}
\keyword{internal}
