% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_shape}
\alias{test_shape}
\title{A sample shapefile}
\format{
A shapefile with a single polygon.
}
\usage{
test_shape
}
\description{
An sf shapefile to show as an example of user-defined GeoJSON in get_event()
and get_raster()
}
\keyword{datasets}
